pyproject.toml
[tool.poetry]
name = "audexa-backend"
version = "0.1.0"
description = "FastAPI backend for Audexa AI"
authors = ["Audexa AI"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic-settings = "^2.6.0"
sqlalchemy = "^2.0.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.0"}
alembic = "^1.13.0"
email-validator = "^2.3.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn main:app --reload"

